git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@9958 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/magick-type.h
@@ -86,9 +86,9 @@ typedef unsigned int Quantum;
 #define MaxColormapSize  65536UL
 #define MaxMap  65535UL
 
-typedef float MagickRealType;
-typedef ssize_t SignedQuantum;
-typedef float Quantum;
+typedef double MagickRealType;
+typedef double SignedQuantum;
+typedef double Quantum;
 #define QuantumRange  65535.0
 #define QuantumFormat  "%g"
 #else
